Release V0.5.1

+++ src/Core/Parallel/Parallel.php
@@ -44,21 +44,26 @@ use ReflectionClass;
 use Revolt\EventLoop;
 use Throwable;
 
+use RuntimeException;
+
+use function Co\cancel;
+use function Co\defer;
+use function Co\onSignal;
 use function count;
 use function dirname;
 use function file_exists;
+use function getmypid;
 use function intval;
 use function is_int;
-use function P\cancel;
-use function P\defer;
-use function P\onSignal;
 use function posix_getpid;
 use function posix_kill;
 use function preg_match;
 use function shell_exec;
+
 use function strval;
 
 use const SIGUSR2;
+use const PHP_OS_FAMILY;
 
 /**
  * 2024-08-07
@@ -126,6 +131,24 @@ class Parallel extends LibraryAbstract
     }
 
     /**
+     * @Author cclilshy
+     * @Date   2024/8/30 22:11
+     * @return static
+     * @throws RuntimeException
+     */
+    public static function getInstance(): static
+    {
+        /**
+         * @compatible:Windows
+         */
+        if (PHP_OS_FAMILY === 'Windows') {
+            throw new RuntimeException('Parallel is not supported on Windows');
+        }
+
+        return parent::getInstance();
+    }
+
+    /**
      * @return void
      */
     private function initialize(): void
@@ -174,11 +197,24 @@ class Parallel extends LibraryAbstract
         $this->counterRuntime = new Runtime();
         $this->counterFuture = $this->counterRuntime->run(static function ($channel, $eventScalar) {
             $eventScalar(fn () => $eventScalar->wait());
-            $processId = posix_getpid();
+            /**
+             * @compatible:Windows
+             */
+            if (PHP_OS_FAMILY === 'Windows') {
+                $processId = getmypid();
+            } else {
+                $processId = posix_getpid();
+            }
             $count = 0;
             while($number = $channel->recv()) {
                 $eventScalar->set($count += $number);
                 if($number > 0) {
+                    /**
+                     * @compatible:Windows
+                     */
+                    if (PHP_OS_FAMILY === 'Windows') {
+                        break;
+                    }
                     posix_kill($processId, SIGUSR2);
                 } elseif($count === -1) {
                     break;
@@ -258,7 +294,7 @@ class Parallel extends LibraryAbstract
      * @param int|null $capacity
      * @return Channel
      */
-    public function makeChannel(string $name, ?int $capacity = null): Channel
+    public function makeChannel(string $name, int|null $capacity = null): Channel
     {
         return is_int($capacity)
             ? new Channel(\parallel\Channel::make($name, $capacity))
